requirements.txt
flask
fpdf
pillow
pandas
